# hconstr don't use push_unknown_rel for matches