I tried this in the original hconstr PR and it didn't seem very impactful, but maybe it was hidden by the main change's gains.

It does seem to help on https://github.com/mit-plv/rewriter/blob/edcec730f68469475fdc4b78495ae941a5b320ec/src/Rewriter/Language/Wf.v#L1341 which appears to have lots of matches.

It should also help if we ever figure out how to exploit preexisting sharing as physically equal terms still need to be discriminated by their free rels, so using push_unknown_rel appears as extra collisions.