@SkySkimmer 2024-07-26
local bench to have a look until the ci machines are freed: (NB my cpu has some hybrid stuff which makes the bench not understand instruction counts)
~~~
┌──────────────┬───────────────────────┬─────────────────┬──────────────────┬──────────────────────────┐
│              │     user time [s]     │   CPU cycles    │ CPU instructions │  max resident mem [KB]   │
│              │                       │                 │                  │                          │
│ package_name │  NEW     OLD    PDIFF │ NEW  OLD  PDIFF │ NEW  OLD  PDIFF  │   NEW      OLD    PDIFF  │
├──────────────┼───────────────────────┼─────────────────┼──────────────────┼──────────────────────────┤
│   coq-stdlib │ 369.64  391.95  -5.69 │   0    0    nan │    0    0    nan │  635848   646588   -1.66 │
│ coq-rewriter │ 284.09  291.36  -2.50 │   0    0    nan │    0    0    nan │ 1474744  1968616  -25.09 │
│     coq-core │ 155.07  155.30  -0.15 │   0    0    nan │    0    0    nan │  467512   468380   -0.19 │
└──────────────┴───────────────────────┴─────────────────┴──────────────────┴──────────────────────────┘
~~~

line changes:
~~~
┌────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────┐
│                                             Significant line time changes in bench                                             │
│                                                                                                                                │
│   OLD      NEW     DIFF      %DIFF     Ln                  FILE                                                                │
├────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────┤
│  7.5960   3.9240  -3.6720    -48.34%  1408  coq-rewriter/src/Rewriter/Language/Wf.v.html                                       │
│  7.6940   6.8610  -0.8330    -10.83%    79  coq-rewriter/src/Rewriter/Rewriter/Examples/PerfTesting/LiftLetsMap.v.html         │
│  1.4730   0.9940  -0.4790    -32.52%   233  coq-rewriter/src/Rewriter/Rewriter/InterpProofs.v.html                             │
│  1.0730   0.6470  -0.4260    -39.70%  1341  coq-rewriter/src/Rewriter/Language/Wf.v.html                                       │
(after this it looks like noise including the slowdowns)
~~~

rewriter NEW raw perf output:
~~~
# started on Fri Jul 26 14:43:26 2024


 Performance counter stats for 'make -j1':

    71,791,667,042      cpu_atom/instructions:u/         #    1.53  insn per cycle              (10.00%)
 2,637,701,559,672      cpu_core/instructions:u/         #   56.20  insn per cycle              (99.33%)
    46,936,871,911      cpu_atom/cycles:u/                                                      (10.00%)
 1,299,974,076,559      cpu_core/cycles:u/                                                      (99.33%)

     291.929136396 seconds time elapsed

     284.085568000 seconds user
       7.758016000 seconds sys
~~~

OLD:
~~~
# started on Fri Jul 26 14:48:53 2024


 Performance counter stats for 'make -j1':

   180,904,228,080      cpu_atom/instructions:u/         #    1.57  insn per cycle              (4.90%)
 2,688,228,109,287      cpu_core/instructions:u/         #   23.34  insn per cycle              (99.24%)
   115,181,751,559      cpu_atom/cycles:u/                                                      (4.90%)
 1,326,300,903,262      cpu_core/cycles:u/                                                      (99.24%)

     298.901881860 seconds time elapsed

     291.360267000 seconds user
       7.666759000 seconds sys
~~~