@SkySkimmer 2024-08-01
@coqbot run full ci